Sergio-VQG/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
h5py==3.1.0
numpy==1.19.5
tensorflow-cpu==2.5.0
